--- frontend/src/App.css
@@ -1,15 +1,0 @@
-.App {
-  /* height: 500px;
-  width: 500px; */
-}
-
-
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
