+++ frontend/src/components/Login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import style from '../Login.module.css'
+import style from '../StyleSheets/Login.module.css'
 class Login extends React.Component{
     constructor(props) {
         super(props);
