+++ frontend/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './StyleSheets/index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import store from './store';
